Move notification schemas and modules into central datastore module
Part of centralizing migrations
